pull out blog post footer
