+++ themes/gatsby-theme-blog/src/components/home-footer-content.js
